+++ SwiftyCoreDataExample/SwiftyCoreDataExample/Persistance/Entities/CatEntity.swift
@@ -19,7 +19,7 @@ public class CatEntity: NSManagedObject {
     
     @NSManaged public var name: String?
     @NSManaged public var weight: Double
-    @NSManaged public var age: Int16
+    @NSManaged public var age: Int64
 
 }
